rasberry_software/requirements.txt
requests>=2.31.0
librosa>=0.10.0
scipy>=1.11.0
numpy>=1.24.0
soundfile>=0.12.0
sounddevice>=0.4.6
wavio>=0.0.8
matplotlib>=3.7.0
RPi.GPIO>=0.7.1; platform_machine=="armv7l" or platform_machine=="aarch64"